app-engine-login-required/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
appengine-python-standard==1.0.0
    # via -r requirements.in
attrs==22.1.0
    # via appengine-python-standard
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via flask
flask==2.2.1
    # via -r requirements.in
frozendict==2.3.4
    # via appengine-python-standard
google-api-core[grpc]==2.8.2
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.10.0
    # via
    #   appengine-python-standard
    #   google-api-core
    #   google-cloud-core
google-cloud-appengine-logging==1.1.3
    # via google-cloud-logging
google-cloud-audit-log==0.2.3
    # via google-cloud-logging
google-cloud-core==2.3.2
    # via google-cloud-logging
google-cloud-logging==3.2.1
    # via -r requirements.in
googleapis-common-protos[grpc]==1.56.4
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.4
    # via google-cloud-logging
grpcio==1.47.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.47.0
    # via google-api-core
idna==3.3
    # via requests
importlib-metadata==4.12.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mock==4.0.3
    # via appengine-python-standard
pillow==9.2.0
    # via appengine-python-standard
proto-plus==1.20.6
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==3.20.1
    # via
    #   appengine-python-standard
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pytz==2022.1
    # via appengine-python-standard
requests==2.28.1
    # via
    #   appengine-python-standard
    #   google-api-core
rsa==4.9
    # via google-auth
ruamel-yaml==0.17.21
    # via appengine-python-standard
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
six==1.16.0
    # via
    #   appengine-python-standard
    #   google-auth
    #   grpcio
urllib3==1.26.11
    # via
    #   appengine-python-standard
    #   requests
werkzeug==2.2.1
    # via flask
zipp==3.8.1
    # via importlib-metadata
